requirements.txt
PostgreSQL psycopg
djangorestframework
django
django-cors-headers
djangorestframework_simplejwt